fix(demo): wasmPitch must be initialized before starting
