+++ demo/demo.html
@@ -7,4 +7,5 @@
         <span id="pitch">__</span> Hz
     </body>
     <script src="bundle.js"></script>
+    <!-- <script src="wasm_interface.js"></script> -->
 </html>
